backend/requirements.txt
# Core Data Processing
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=12.0.0  # For parquet file support

# Pattern Matching & ML
faiss-cpu>=1.7.4  # Facebook AI Similarity Search
scikit-learn>=1.3.0
scipy>=1.11.0
dtw-python>=1.3.0  # Dynamic Time Warping

# Data Sources
yfinance>=0.2.28  # Yahoo Finance data
ccxt>=4.0.0  # Crypto exchange data
requests>=2.31.0

# Firebase
firebase-admin>=6.2.0

# API Server
fastapi>=0.103.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.4.0
gunicorn>=21.0.0  # Production WSGI server

# Async Support
aiohttp>=3.8.5
asyncio>=3.4.3

# Trading - MetaTrader 5 (optional, Windows only)
# MetaTrader5>=5.0.45

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.2
